Node state properly readback to scene
Fixed an issue where the graph was incrementing through even when nodes were not active. This was due to checking a value against the list of nodes instead of just checking to see if a node had started and if it was running. This should allow for future implementation into other scene based node variables